+++ libs/@blockprotocol/hook/src/hook-embedder-handler.ts
@@ -1,27 +1,27 @@
-import { ServiceHandler } from "@blockprotocol/core";
+import { ModuleHandler } from "@blockprotocol/core";
 
 // @todo restore this when an issue with module resolution has been resolved
-// import hookServiceJson from "./hook-service.json" assert { type: "json" };
-import { EmbedderHookMessageCallbacks, EmbedderHookMessages } from "./types.js";
+// import hookModuleJson from "./hook-module.json" assert { type: "json" };
+import { HookEmbedderMessageCallbacks, HookEmbedderMessages } from "./types.js";
 
 /**
- * Creates a handler for the hook service for the embedder.
+ * Creates a handler for the hook module for the embedder.
  * Register callbacks in the constructor or afterwards using the 'on' method to
  * react to messages from the block. Call the relevant methods to send messages
  * to the block.
  */
 export class HookEmbedderHandler
-  extends ServiceHandler
-  implements EmbedderHookMessages
+  extends ModuleHandler
+  implements HookEmbedderMessages
 {
   constructor({
     callbacks,
     element,
   }: {
-    callbacks?: Partial<EmbedderHookMessageCallbacks>;
+    callbacks?: Partial<HookEmbedderMessageCallbacks>;
     element?: HTMLElement | null;
   }) {
-    super({ element, callbacks, serviceName: "hook", sourceType: "embedder" });
+    super({ element, callbacks, moduleName: "hook", sourceType: "embedder" });
   }
 
   /**
@@ -33,15 +33,15 @@ export class HookEmbedderHandler
    * @param handlerFunction the function to call when the message is received,
    *   with the message data / errors
    */
-  on<K extends keyof EmbedderHookMessageCallbacks>(
+  on<K extends keyof HookEmbedderMessageCallbacks>(
     this: HookEmbedderHandler,
     messageName: K,
-    handlerFunction: NonNullable<EmbedderHookMessageCallbacks[K]>,
+    handlerFunction: NonNullable<HookEmbedderMessageCallbacks[K]>,
   ) {
     // @todo restore this when module resolution issue resolved
     // @see https://app.asana.com/0/1202542409311090/1202614421149286/f
     // const expectedMessageSource = "block";
-    // const messageJsonDefinition = hookServiceJson.messages.find(
+    // const messageJsonDefinition = hookModuleJson.messages.find(
     //   (message) =>
     //     message.messageName === messageName &&
     //     message.source === expectedMessageSource,